Convert encounters module
